OE0-68: renamed mapping file
